+++ R/all_list_elements.R
@@ -23,13 +23,13 @@ is_fully_named_list <- function(x){
 #' Are all elements in a list null or data.frames?
 #' @param x An object
 #' @examples
-#' all_list_elements_null_or_df(data.frame())
-#' all_list_elements_null_or_df(list(data.frame()))
-#' all_list_elements_null_or_df(list(1, NULL))
-#' all_list_elements_null_or_df(list(data.frame(), NULL))
-#' all_list_elements_null_or_df(list("a"=1, 2))
+#' is_all_list_elements_null_or_df(data.frame())
+#' is_all_list_elements_null_or_df(list(data.frame()))
+#' is_all_list_elements_null_or_df(list(1, NULL))
+#' is_all_list_elements_null_or_df(list(data.frame(), NULL))
+#' is_all_list_elements_null_or_df(list("a"=1, 2))
 #' @export
-all_list_elements_null_or_df <- function(x){
+is_all_list_elements_null_or_df <- function(x){
   if(!inherits(x, "list")) return(FALSE)
 
   retval <- lapply(x, function(x) !inherits(x, c("data.frame", "NULL")))
@@ -42,13 +42,13 @@ all_list_elements_null_or_df <- function(x){
 #' Are all elements in a list null or lists?
 #' @param x An object
 #' @examples
-#' all_list_elements_null_or_list(data.frame())
-#' all_list_elements_null_or_list(list(data.frame()))
-#' all_list_elements_null_or_list(list(1, NULL))
-#' all_list_elements_null_or_list(list(list(), NULL))
-#' all_list_elements_null_or_list(list("a"=1, 2))
+#' is_all_list_elements_null_or_list(data.frame())
+#' is_all_list_elements_null_or_list(list(data.frame()))
+#' is_all_list_elements_null_or_list(list(1, NULL))
+#' is_all_list_elements_null_or_list(list(list(), NULL))
+#' is_all_list_elements_null_or_list(list("a"=1, 2))
 #' @export
-all_list_elements_null_or_list <- function(x){
+is_all_list_elements_null_or_list <- function(x){
   if(!inherits(x, "list")) return(FALSE)
 
   retval <- lapply(x, function(x) !inherits(x, c("list", "NULL")))
@@ -61,14 +61,14 @@ all_list_elements_null_or_list <- function(x){
 #' Are all elements in a list null or fully named lists?
 #' @param x An object
 #' @examples
-#' all_list_elements_null_or_fully_named_list(data.frame())
-#' all_list_elements_null_or_fully_named_list(list(data.frame()))
-#' all_list_elements_null_or_fully_named_list(list(1, NULL))
-#' all_list_elements_null_or_fully_named_list(list(list(), NULL))
-#' all_list_elements_null_or_fully_named_list(list(list("a" = 1), NULL))
-#' all_list_elements_null_or_fully_named_list(list("a"=1, 2))
+#' is_all_list_elements_null_or_fully_named_list(data.frame())
+#' is_all_list_elements_null_or_fully_named_list(list(data.frame()))
+#' is_all_list_elements_null_or_fully_named_list(list(1, NULL))
+#' is_all_list_elements_null_or_fully_named_list(list(list(), NULL))
+#' is_all_list_elements_null_or_fully_named_list(list(list("a" = 1), NULL))
+#' is_all_list_elements_null_or_fully_named_list(list("a"=1, 2))
 #' @export
-all_list_elements_null_or_fully_named_list <- function(x){
+is_all_list_elements_null_or_fully_named_list <- function(x){
   if(!inherits(x, "list")) return(FALSE)
 
   are_any_lists_unnamed <- lapply(x, function(y){
@@ -88,7 +88,7 @@ all_list_elements_null_or_fully_named_list <- function(x){
 
   are_lists_named <- !are_any_lists_unnamed
 
-  retval <- all_list_elements_null_or_list(x) & are_lists_named
+  retval <- is_all_list_elements_null_or_list(x) & are_lists_named
 
   return(retval)
 }